Add HttpStatus Codes

+++ CinemaApi/Controllers/MoviesController.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using CinemaApi.Data;
 using CinemaApi.Models;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -22,47 +23,54 @@ namespace CinemaApi.Controllers
         }
         // GET: api/<MoviesController>
         [HttpGet]
-        public IEnumerable<Movie> Get()
+        public IActionResult Get()
         {
-            return _dbContext.Movies;
+            return Ok(_dbContext.Movies);
         }
 
         // GET api/<MoviesController>/5
         [HttpGet("{id}")]
-        public Movie Get(int id)
+        public IActionResult Get(int id)
         {
             var movieInDB = _dbContext.Movies.Find(id);
-            return movieInDB;
+            if (movieInDB == null)
+                return NotFound("No record found");
+            return Ok(movieInDB);
         }
 
         // POST api/<MoviesController>
         [HttpPost]
-        public Movie Post([FromBody] Movie movie)
+        public IActionResult Post([FromBody] Movie movie)
         {
             _dbContext.Movies.Add(movie);
             _dbContext.SaveChanges();
-            return movie;
+            return StatusCode(StatusCodes.Status201Created);
         }
 
         // PUT api/<MoviesController>/5
         [HttpPut("{id}")]
-        public Movie Put(int id, [FromBody] Movie movie)
+        public IActionResult Put(int id, [FromBody] Movie movie)
         {
             var movieInDb = _dbContext.Movies.Find(id);
+            if (movieInDb == null)
+                return NotFound("No Record found of the same ID");
             movieInDb.Name = movie.Name;
             movieInDb.Language = movie.Language;
             _dbContext.SaveChanges();
-            return movieInDb;
+            return Ok("Record Updated Successfully");
             
         }
 
         // DELETE api/<MoviesController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
             var movieInDb = _dbContext.Movies.Find(id);
+            if (movieInDb == null)
+                return NotFound("No Record found of the same ID");
             _dbContext.Movies.Remove(movieInDb);
             _dbContext.SaveChanges();
+            return Ok("Record Deleted");
         }
     }
 }
